Refactored the way local modules get loaded
Instead of files getting loaded via the normal require()
now they get loaded through a custom function so that
the paths are not duplicated all over the place. Now it
has more of a "package" feel where you include depedencies based
on the package they are on. For example "logger/logger"